* master: (21 commits)
  Added missing import statement in metrics.py
  Isolated label fix (#187)
  Track all embeddin outputs in Snakefile
  Set key name when computing graph layout
  including knn_graph files in package installation
  solved typo that broke finding optimal score for clustering
  Add embedding stage to Snakemake
  Revert metrics.py to old version
  More embedding plots to new save_embeddings script
  Fix small bug in scanvi function
  Update runMethods.R
  Remove _ in "seurat_rpca" to facilitate parsing
  added seurat rpca
  added new methods names
  Fix adata references
  Create output directory for figures
  Fix method name check
  Output embedding coordinates as CSV
  Use graph embedding for CONOS when saving figures
  Tidy comments
  ...

+++ scIB/integration.py
@@ -227,7 +227,7 @@ def runScanvi(adata, batch, labels):
 
     # # Defaults from SCVI github tutorials scanpy_pbmc3k and harmonization
     n_epochs_scVI = np.min([round((20000/adata.n_obs)*400), 400]) #400
-    n_epochs_scANVI = np.min([10, np.max([2, round(n_epochs_scVI / 3.)])])
+    n_epochs_scANVI = int(np.min([10, np.max([2, round(n_epochs_scVI / 3.)])]))
     n_latent=30
     n_hidden=128
     n_layers=2
